Make view models platform-independent and fix tests
