+++ CombineStubs.swift
@@ -1,0 +1,7 @@
+#if !canImport(Combine)
+public protocol ObservableObject {}
+@propertyWrapper public struct Published<Value> {
+    public var wrappedValue: Value
+    public init(wrappedValue: Value) { self.wrappedValue = wrappedValue }
+}
+#endif
